usp_job_stream_status
Proc created in an attempt to replicate data displayed in Job Stream Status and Job Status History tabs of the Paragon Job Stream module. It's scripted to use in an Agent Job to alert an operator or user at any given time on the history or current status of jobs.